fix: cambiar IDs a Integer en DTOs y corregir extensión de servicio
- Se actualizaron los tipos de datos de los IDs (personaId, metodoPagoId, inventarioId y ventaId) de Long a Integer en los DTOs de Compra para alinearlos con la base de datos y las entidades JPA.
- Se corrigió la extensión de CompraServiceImpl.java a minúscula para evitar inconvenientes en entornos Linux.

+++ backend/src/main/java/com/tienda/backend/dto/CompraDetalleDTO.java
@@ -1,23 +1,23 @@
 package com.tienda.backend.dto;
 
 public class CompraDetalleDTO {
-    private Long inventarioId;
+    private Integer inventarioId;
     private Integer cantidad;
 
     // Constructores
     public CompraDetalleDTO() {}
 
-    public CompraDetalleDTO(Long inventarioId, Integer cantidad) {
+    public CompraDetalleDTO(Integer inventarioId, Integer cantidad) {
         this.inventarioId = inventarioId;
         this.cantidad = cantidad;
     }
 
     // Getters y Setters
-    public Long getInventarioId() {
+    public Integer getInventarioId() {
         return inventarioId;
     }
 
-    public void setInventarioId(Long inventarioId) {
+    public void setInventarioId(Integer inventarioId) {
         this.inventarioId = inventarioId;
     }
